+++ CHANGELOG.md
@@ -6,12 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+- `bill`: normalize line discounts and charges to remove empty rows.
+
+### Fixed
+
+- `tax`: identity code handling will skip default validation for specific countries that use special characters.
+
 ## [v0.204.0]
 
 ### Added
 
 - `br-nfse-v1`: added initial Brazil NFS-e addon
-- `bill`: normalize line discounts and charges to remove empty rows.
 
 ### Changed